% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_wald_ci}
\alias{calculate_wald_ci}
\title{Confidence Interval(Wald method)}
\usage{
calculate_wald_ci(alpha, n, p_hat)
}
\arguments{
\item{alpha}{Type I error}

\item{n}{Sample size of the data}

\item{p_hat}{Estimated value of Evaluation Metrics}
}
\value{
A list of confidence interval(lower/upper)
}
\description{
This function calculate Confidence Interval using wald method
}
