% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data}
\format{
A data frame with N(default=300) rows and 3 variables:
\describe{
\item{actual}{the true class label of each observation}
\item{predicted}{the predicted class label of each observation}
\item{score}{the predicted score for each observation}
}
}
\usage{
data("example_data")
}
\description{
This data is used for testing the other codes
}
\keyword{datasets}
