% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_format.R
\name{format_performance_output}
\alias{format_performance_output}
\title{Format performance metrics into a data frame}
\usage{
format_performance_output(metrics)
}
\arguments{
\item{metrics}{A list of metric results returned by calc_performance_metrics}
}
\value{
A data.frame with columns: Metric, Estimate, Lower_95_CI, Upper_95_CI
}
\description{
Format performance metrics into a data frame
}
